@@ -29,21 +29,21 @@ setup(
         "Operating System :: Microsoft",
         "Programming Language :: Python",
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.5",
         "Programming Language :: Python :: 3.6",
         "Programming Language :: Python :: 3.7",
         "Programming Language :: Python :: 3.8",
+        "Programming Language :: Python :: 3.9",
+        "Programming Language :: Python :: 3.10",
         "Programming Language :: Python :: Implementation :: PyPy",
         "Programming Language :: Python :: Implementation :: CPython",
     ],
     packages=find_packages(exclude=["tests"]),
     package_data={"aio_pika": ["py.typed"]},
-    install_requires=["aiormq>=3.2.3,<4", "yarl"],
-    python_requires=">=3.5, <4",
+    install_requires=["aiormq~=6.2.2", "yarl"],
+    python_requires=">3.6, <4",
     extras_require={
         "develop": [
-            "aiomisc~=10.1.6",
-            "async_generator",
+            "aiomisc~=15.6.8",
             "coverage!=4.3",
             "coveralls",
             "pylava",
