@@ -1,114 +1,127 @@
 import asyncio
 import logging
 from functools import partial
-from typing import Optional, Type, TypeVar
+from types import TracebackType
+from typing import Any, Callable, Dict, Optional, Tuple, Type, TypeVar, Union
+from warnings import warn
 
 import aiormq
+from aiormq.abc import ExceptionType
 from aiormq.tools import censor_url
+from pamqp.common import FieldTable
 from yarl import URL
 
+# This needed only for migration from 6.x to 7.x
+# TODO: Remove this in 8.x release
+from .abc import ConnectionType  # noqa
+from .abc import (
+    AbstractChannel, AbstractConnection, ConnectionCloseCallback, TimeoutType,
+)
 from .channel import Channel
-from .pool import PoolInstance
 from .tools import CallbackCollection
-from .types import CloseCallbackType, TimeoutType
 
 
 log = logging.getLogger(__name__)
+T = TypeVar("T")
 
 
-class Connection(PoolInstance):
+class Connection(AbstractConnection):
     """ Connection abstraction """
 
-    CHANNEL_CLASS = Channel
-    KWARGS_TYPES = ()
+    CHANNEL_CLASS: Type[Channel] = Channel
+    KWARGS_TYPES: Tuple[Tuple[str, Callable[[str], Any], str], ...] = ()
 
     @property
-    def is_closed(self):
+    def is_closed(self) -> bool:
         return self.closing.done()
 
-    async def close(self, exc=asyncio.CancelledError):
+    async def close(
+        self, exc: Optional[ExceptionType] = asyncio.CancelledError,
+    ) -> None:
         if not self.closing.done():
             self.closing.set_result(exc)
 
-        return await self.connection.close(exc)
+        if not hasattr(self, "connection"):
+            return None
+
+        await self.connection.close(exc)
+
+        if hasattr(self, "connection"):
+            del self.connection
 
     @classmethod
-    def _parse_kwargs(cls, kwargs):
+    def _parse_kwargs(cls, kwargs: Dict[str, Any]) -> Dict[str, Any]:
         result = {}
         for key, parser, default in cls.KWARGS_TYPES:
             result[key] = parser(kwargs.get(key, default))
         return result
 
     def __init__(
-        self, url, loop: Optional[asyncio.AbstractEventLoop] = None, **kwargs
+        self, url: URL, loop: Optional[asyncio.AbstractEventLoop] = None,
+        **kwargs: Any
     ):
+        super().__init__(url, loop, **kwargs)
+
         self.loop = loop or asyncio.get_event_loop()
         self.url = URL(url)
 
-        self.kwargs = self._parse_kwargs(kwargs or self.url.query)
+        self.kwargs: Dict[str, Any] = self._parse_kwargs(
+            kwargs or dict(self.url.query),
+        )
 
-        self._close_callbacks = CallbackCollection(self)
-        self.connection = None  # type: Optional[aiormq.Connection]
-        self.closing = self.loop.create_future()
+        self.connection: aiormq.abc.AbstractConnection
+        self.close_callbacks = CallbackCollection(self)
+        self.connected: asyncio.Event = asyncio.Event()
+        self.closing: asyncio.Future = self.loop.create_future()
 
-    @property
-    def close_callbacks(self) -> CallbackCollection:
-        return self._close_callbacks
-
-    @property
-    def heartbeat_last(self) -> float:
-        """ returns loop.time() value since last received heartbeat """
-        return self.connection.heartbeat_last_received
-
-    @property
-    def _channels(self) -> dict:
-        return self.connection.channels
-
-    def __str__(self):
+    def __str__(self) -> str:
         return str(censor_url(self.url))
 
-    def __repr__(self):
-        return '<{0}: "{1}">'.format(self.__class__.__name__, str(self))
+    def __repr__(self) -> str:
+        return f'<{self.__class__.__name__}: "{self}">'
 
     def add_close_callback(
-        self, callback: CloseCallbackType, weak: bool = False
-    ):
-        """ Add callback which will be called after connection will be closed.
-
-        :class:`BaseException` or None will be passed as a first argument.
-
-        Example:
-
-        .. code-block:: python
-
-            import aio_pika
-
-            async def main():
-                connection = await aio_pika.connect(
-                    "amqp://guest:guest@127.0.0.1/"
-                )
-                connection.add_close_callback(print)
-                await connection.close()
-                # None
-
-
-        :return: None
-        """
+        self, callback: ConnectionCloseCallback, weak: bool = False,
+    ) -> None:
+        warn(
+            "This method will be removed from future release. "
+            f"Use {self.__class__.__name__}.close_callbacks.add instead",
+            DeprecationWarning,
+            stacklevel=2,
+        )
         self.close_callbacks.add(callback, weak=weak)
 
-    def _on_connection_close(self, connection, closing, *args, **kwargs):
-        exc = closing.exception()
-        self.close_callbacks(exc)
+    def _on_connection_close(
+        self, connection: AbstractConnection, closing: asyncio.Future,
+    ) -> None:
         log.debug("Closing AMQP connection %r", connection)
+        exc: Optional[BaseException] = closing.exception()
+        self.close_callbacks(exc)
 
-    async def _make_connection(self, **kwargs) -> aiormq.Connection:
-        connection = await aiormq.connect(self.url, **kwargs)
-        connection.closing.add_done_callback(
-            partial(self._on_connection_close, self.connection),
+        if self.closing.done():
+            return
+
+        if exc is not None:
+            self.closing.set_exception(exc)
+            return
+
+        self.closing.set_result(closing.result())
+
+    async def _make_connection(
+        self, *, timeout: TimeoutType = None, **kwargs: Any
+    ) -> aiormq.abc.AbstractConnection:
+        connection: aiormq.abc.AbstractConnection = await asyncio.wait_for(
+            aiormq.connect(self.url, **kwargs), timeout=timeout,
         )
+        connection.closing.add_done_callback(
+            partial(self._on_connection_close, connection),
+        )
+        await connection.ready()
         return connection
 
-    async def connect(self, timeout: TimeoutType = None, **kwargs):
+    async def connect(
+        self, timeout: TimeoutType = None, **kwargs: Any
+    ) -> None:
         """ Connect to AMQP server. This method should be called after
         :func:`aio_pika.connection.Connection.__init__`
 
@@ -117,8 +130,12 @@ class Connection(PoolInstance):
             You shouldn't call it explicitly.
 
         """
-        self.connection = await asyncio.wait_for(
-            self._make_connection(**kwargs), timeout=timeout,
+        self.connection = (
+            await self._make_connection(timeout=timeout, **kwargs)
+        )
+        self.connected.set()
+        self.connection.closing.add_done_callback(
+            lambda _: self.connected.clear(),
         )
 
     def channel(
@@ -126,7 +143,7 @@ class Connection(PoolInstance):
         channel_number: int = None,
         publisher_confirms: bool = True,
         on_return_raises: bool = False,
-    ) -> Channel:
+    ) -> AbstractChannel:
         """ Coroutine which returns new instance of :class:`Channel`.
 
         Example:
@@ -192,12 +209,15 @@ class Connection(PoolInstance):
         log.debug("Channel created: %r", channel)
         return channel
 
-    async def ready(self):
-        while self.connection is None:
-            await asyncio.sleep(0)
+    async def ready(self) -> None:
+        await self.connected.wait()
 
-    def __del__(self):
-        if any((self.is_closed, self.loop.is_closed(), not self.connection)):
+    def __del__(self) -> None:
+        if (
+            self.is_closed or
+            self.loop.is_closed() or
+            not hasattr(self, "connection")
+        ):
             return
 
         asyncio.shield(self.close())
@@ -205,19 +225,49 @@ class Connection(PoolInstance):
     async def __aenter__(self) -> "Connection":
         return self
 
-    async def __aexit__(self, exc_type, exc_val, exc_tb):
-        for channel in tuple(self._channels.values()):  # can change size
-            if channel and not channel.is_closed:
-                await channel.close()
-
+    async def __aexit__(
+        self,
+        exc_type: Optional[Type[BaseException]],
+        exc_val: Optional[BaseException],
+        exc_tb: Optional[TracebackType],
+    ) -> None:
         await self.close()
 
 
-ConnectionType = TypeVar("ConnectionType", bound=Connection)
+def make_url(
+    url: Union[str, URL] = None,
+    *,
+    host: str = "localhost",
+    port: int = 5672,
+    login: str = "guest",
+    password: str = "guest",
+    virtualhost: str = "/",
+    ssl: bool = False,
+    ssl_options: dict = None,
+    **kwargs: Any
+) -> URL:
+    if url is not None:
+        if not isinstance(url, URL):
+            return URL(url)
+        return url
+
+    kw = kwargs
+    kw.update(ssl_options or {})
+
+    return URL.build(
+        scheme="amqps" if ssl else "amqp",
+        host=host,
+        port=port,
+        user=login,
+        password=password,
+        # yarl >= 1.3.0 requires path beginning with slash
+        path="/" + virtualhost,
+        query=kw,
+    )
 
 
 async def connect(
-    url: str = None,
+    url: Union[str, URL] = None,
     *,
     host: str = "localhost",
     port: int = 5672,
@@ -228,10 +278,10 @@ async def connect(
     loop: asyncio.AbstractEventLoop = None,
     ssl_options: dict = None,
     timeout: TimeoutType = None,
-    connection_class: Type[ConnectionType] = Connection,
-    client_properties: dict = None,
-    **kwargs
-) -> ConnectionType:
+    client_properties: FieldTable = None,
+    connection_class: Type[AbstractConnection] = Connection,
+    **kwargs: Any
+) -> AbstractConnection:
 
     """ Make connection to the broker.
 
@@ -313,26 +363,27 @@ async def connect(
 
     """
 
-    if url is None:
-        kw = kwargs
-        kw.update(ssl_options or {})
-
-        url = URL.build(
-            scheme="amqps" if ssl else "amqp",
+    connection: AbstractConnection = connection_class(
+        make_url(
+            url,
             host=host,
             port=port,
-            user=login,
+            login=login,
             password=password,
-            # yarl >= 1.3.0 requires path beginning with slash
-            path="/" + virtualhost,
-            query=kw,
-        )
-
-    connection = connection_class(url, loop=loop)
+            virtualhost=virtualhost,
+            ssl=ssl,
+            ssl_options=ssl_options,
+            **kwargs
+        ),
+        loop=loop,
+    )
 
     await connection.connect(
-        timeout=timeout, client_properties=client_properties, loop=loop,
+        timeout=timeout,
+        client_properties=client_properties,
+        loop=loop,
     )
+
     return connection
 
 
