@@ -1,39 +1,39 @@
 import asyncio
-from enum import Enum, unique
 from logging import getLogger
-from typing import Optional, Union
+from types import TracebackType
+from typing import Any, Awaitable, Generator, Optional, Type, Union
 from warnings import warn
 
 import aiormq
-import aiormq.types
+import aiormq.abc
+from pamqp.common import Arguments
 
+from .abc import (
+    AbstractChannel, AbstractConnection, AbstractExchange, AbstractQueue,
+    ChannelCloseCallback, TimeoutType,
+)
 from .exchange import Exchange, ExchangeType
+from .message import ReturnCallback  # noqa
 from .message import IncomingMessage
-from .pool import PoolInstance
 from .queue import Queue
-from .tools import CallbackCollection
+from .tools import CallbackCollection, task
 from .transaction import Transaction
-from .types import CloseCallbackType, ReturnCallbackType, TimeoutType
 
 
 log = getLogger(__name__)
 
 
-@unique
-class ConfirmationTypes(Enum):
-    ACK = "ack"
-    NACK = "nack"
-
-
-class Channel(PoolInstance):
+class Channel(AbstractChannel):
     """ Channel abstraction """
 
     QUEUE_CLASS = Queue
     EXCHANGE_CLASS = Exchange
 
+    _channel: aiormq.abc.AbstractChannel
+
     def __init__(
         self,
-        connection,
+        connection: AbstractConnection,
         channel_number: Optional[int] = None,
         publisher_confirms: bool = True,
         on_return_raises: bool = False,
@@ -56,175 +56,208 @@ class Channel(PoolInstance):
 
         self.loop = connection.loop
 
-        self._connection = connection
-        self._done_callbacks = CallbackCollection(self)
-        self._return_callbacks = CallbackCollection(self)
-        self._channel = None  # type: Optional[aiormq.Channel]
+        self._channel: aiormq.abc.AbstractChannel
         self._channel_number = channel_number
+
+        self.connection = connection
+        self.close_callbacks = CallbackCollection(self)
+        self.return_callbacks = CallbackCollection(self)
+
         self._on_return_raises = on_return_raises
         self._publisher_confirms = publisher_confirms
 
         self._delivery_tag = 0
 
-        # noinspection PyTypeChecker
-        self.default_exchange = None  # type: Optional[Exchange]
+        # That's means user closed channel instance explicitly
+        self._is_closed_by_user: bool = False
+
+        self.default_exchange: Exchange
 
     @property
     def done_callbacks(self) -> CallbackCollection:
-        return self._done_callbacks
+        return self.close_callbacks
 
     @property
-    def return_callbacks(self) -> CallbackCollection:
-        return self._return_callbacks
+    def is_initialized(self) -> bool:
+        """ Returns True when the channel has been opened
+        and ready for interaction """
+        return hasattr(self, "_channel")
 
     @property
-    def is_closed(self):
-        if not self._channel or self._is_closed_by_user:
+    def is_closed(self) -> bool:
+        """ Returns True when the channel has been closed from the broker
+        side or after the close() method has been called. """
+        if not self.is_initialized or self._is_closed_by_user:
             return True
         return self._channel.is_closed
 
-    @property
-    def _is_closed_by_user(self):
-        return self._channel == ()
-
-    @_is_closed_by_user.setter
-    def _is_closed_by_user(self, value: bool):
-        if value:
-            self._channel = ()
-
-    async def close(self, exc=None):
-        if not self._channel:
-            log.warning("Channel already closed")
+    @task
+    async def close(self, exc: aiormq.abc.ExceptionType = None) -> None:
+        if not self.is_initialized:
+            log.warning("Channel not opened")
             return
 
-        if self.is_closed:
-            return
-
-        # noinspection PyTypeChecker
-        channel = self._channel  # type: aiormq.Channel
+        channel: aiormq.abc.AbstractChannel = self._channel
+        del self._channel
         self._is_closed_by_user = True
         await channel.close()
 
-        self._done_callbacks(exc)
-
     @property
-    def channel(self) -> aiormq.Channel:
-        if self._channel is None:
+    def channel(self) -> aiormq.abc.AbstractChannel:
+        if not self.is_initialized:
             raise aiormq.exceptions.ChannelInvalidStateError(
-                "Channel was not opened"
+                "Channel was not opened",
             )
 
         if self.is_closed:
             raise aiormq.exceptions.ChannelInvalidStateError(
-                "Channel has been closed"
+                "Channel has been closed",
             )
 
         return self._channel
 
     @property
-    def number(self):
-        return self._channel.number if self._channel else None
+    def number(self) -> Optional[int]:
+        return self._channel.number if self.is_initialized else None
 
-    def __str__(self):
-        return "{0}".format(self.number or "Not initialized channel")
+    def __str__(self) -> str:
+        return "{}".format(self.number or "Not initialized channel")
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         conn = None
 
-        if self._channel:
+        if self.is_initialized:
             conn = self._channel.connection
 
-        return '<%s "%s#%s">' % (self.__class__.__name__, conn, self)
+        return '<%s #%s "%s">' % (self.__class__.__name__, self, conn)
 
-    def __iter__(self):
-        return (yield from self.__await__())
-
-    def __await__(self):
+    def __await__(self) -> Generator[Any, Any, "AbstractChannel"]:
         yield from self.initialize().__await__()
         return self
 
-    async def __aenter__(self):
-        if not self._channel:
+    async def __aenter__(self) -> "AbstractChannel":
+        if not self.is_initialized:
             await self.initialize()
-
         return self
 
-    async def __aexit__(self, exc_type, exc_val, exc_tb):
-        await self.close()
+    def __aexit__(
+        self,
+        exc_type: Optional[Type[BaseException]],
+        exc_val: Optional[BaseException],
+        exc_tb: Optional[TracebackType],
+    ) -> Awaitable[Any]:
+        return self.close()
 
     def add_close_callback(
-        self, callback: CloseCallbackType, weak: bool = False
+        self, callback: ChannelCloseCallback, weak: bool = False,
     ) -> None:
-        self._done_callbacks.add(callback, weak=weak)
+        warn(
+            "This method will be removed from future release. "
+            f"Use {self.__class__.__name__}.close_callbacks.add instead",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        self.close_callbacks.add(callback, weak=weak)
 
     def remove_close_callback(
-        self, callback: CloseCallbackType
+        self, callback: ChannelCloseCallback,
     ) -> None:
-        self._done_callbacks.remove(callback)
+        warn(
+            "This method will be removed from future release. "
+            f"Use {self.__class__.__name__}.close_callbacks.remove instead",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        self.close_callbacks.remove(callback)
 
     def add_on_return_callback(
-        self, callback: ReturnCallbackType, weak: bool = False,
+        self, callback: ReturnCallback, weak: bool = False,
     ) -> None:
-        self._return_callbacks.add(callback, weak=weak)
+        warn(
+            "This method will be removed from future release. "
+            f"Use {self.__class__.__name__}.return_callbacks.add instead",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        self.return_callbacks.add(callback, weak=weak)
 
-    def remove_on_return_callback(
-        self, callback: ReturnCallbackType, weak: bool = False
-    ) -> None:
-        self._return_callbacks.remove(callback)
+    def remove_on_return_callback(self, callback: ReturnCallback) -> None:
+        warn(
+            "This method will be removed from future release. "
+            f"Use {self.__class__.__name__}.return_callbacks.remove instead",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        self.return_callbacks.remove(callback)
 
-    async def _create_channel(self) -> aiormq.Channel:
-        await self._connection.ready()
+    async def _create_channel(
+        self, timeout: TimeoutType = None,
+    ) -> aiormq.abc.AbstractChannel:
+        await self.connection.ready()
 
-        return await self._connection.connection.channel(
+        return await self.connection.connection.channel(
             publisher_confirms=self._publisher_confirms,
             on_return_raises=self._on_return_raises,
             channel_number=self._channel_number,
+            timeout=timeout,
         )
 
     async def initialize(self, timeout: TimeoutType = None) -> None:
-        if self._channel is not None:
-            raise RuntimeError("Can't initialize channel")
+        if self.is_initialized:
+            raise RuntimeError("Already initialized")
+        elif self._is_closed_by_user:
+            raise RuntimeError("Can't initialize closed channel")
 
-        self._channel = await asyncio.wait_for(
-            self._create_channel(), timeout=timeout,
+        channel: aiormq.abc.AbstractChannel = await self._create_channel(
+            timeout=timeout,
         )
 
+        self._channel = channel
         self._delivery_tag = 0
 
-        if self.default_exchange is None:
-            self.default_exchange = self.EXCHANGE_CLASS(
-                connection=self._connection,
-                channel=self.channel,
-                arguments=None,
-                auto_delete=None,
-                durable=None,
-                internal=None,
-                name="",
-                passive=None,
-                type=ExchangeType.DIRECT,
-            )
+        self.default_exchange = self.EXCHANGE_CLASS(
+            connection=self.connection,
+            channel=self,
+            arguments=None,
+            auto_delete=False,
+            durable=False,
+            internal=False,
+            name="",
+            passive=False,
+            type=ExchangeType.DIRECT,
+        )
 
+        self._on_initialized()
+
+    def _on_channel_closed(self, closing: asyncio.Future) -> None:
+        self.close_callbacks(closing.exception())
+
+    def _on_initialized(self) -> None:
         self.channel.on_return_callbacks.add(self._on_return)
-        self.channel.closing.add_done_callback(self._done_callbacks)
+        self.channel.closing.add_done_callback(self._on_channel_closed)
 
-    def _on_return(self, message: aiormq.types.DeliveredMessage):
-        self._return_callbacks(IncomingMessage(message, no_ack=True))
+    def _on_return(self, message: aiormq.abc.DeliveredMessage) -> None:
+        self.return_callbacks(IncomingMessage(message, no_ack=True))
 
-    async def reopen(self):
-        self._channel = None
+    async def reopen(self) -> None:
+        if hasattr(self, "_channel"):
+            del self._channel
+
+        self._is_closed_by_user = False
         await self.initialize()
 
     async def declare_exchange(
         self,
         name: str,
         type: Union[ExchangeType, str] = ExchangeType.DIRECT,
-        durable: bool = None,
+        *,
+        durable: bool = False,
         auto_delete: bool = False,
         internal: bool = False,
         passive: bool = False,
-        arguments: dict = None,
+        arguments: Arguments = None,
         timeout: TimeoutType = None,
-    ) -> Exchange:
+    ) -> AbstractExchange:
         """
         Declare an exchange.
 
@@ -249,8 +282,8 @@ class Channel(PoolInstance):
             durable = False
 
         exchange = self.EXCHANGE_CLASS(
-            connection=self._connection,
-            channel=self.channel,
+            connection=self.connection,
+            channel=self,
             name=name,
             type=type,
             durable=durable,
@@ -266,7 +299,9 @@ class Channel(PoolInstance):
 
         return exchange
 
-    async def get_exchange(self, name: str, *, ensure: bool = True) -> Exchange:
+    async def get_exchange(
+        self, name: str, *, ensure: bool = True
+    ) -> AbstractExchange:
         """
         With ``ensure=True``, it's a shortcut for
         ``.declare_exchange(..., passive=True)``; otherwise, it returns an
@@ -288,10 +323,10 @@ class Channel(PoolInstance):
             return await self.declare_exchange(name=name, passive=True)
         else:
             return self.EXCHANGE_CLASS(
-                connection=self._connection,
-                channel=self.channel,
+                connection=self.connection,
+                channel=self,
                 name=name,
-                durable=None,
+                durable=False,
                 auto_delete=False,
                 internal=False,
                 passive=True,
@@ -302,13 +337,13 @@ class Channel(PoolInstance):
         self,
         name: str = None,
         *,
-        durable: bool = None,
+        durable: bool = False,
         exclusive: bool = False,
         passive: bool = False,
         auto_delete: bool = False,
-        arguments: dict = None,
+        arguments: Arguments = None,
         timeout: TimeoutType = None
-    ) -> Queue:
+    ) -> AbstractQueue:
         """
 
         :param name: queue name
@@ -328,9 +363,8 @@ class Channel(PoolInstance):
         :raises: :class:`aio_pika.exceptions.ChannelClosed` instance
         """
 
-        queue = self.QUEUE_CLASS(
-            connection=self,
-            channel=self.channel,
+        queue: AbstractQueue = self.QUEUE_CLASS(
+            channel=self,
             name=name,
             durable=durable,
             exclusive=exclusive,
@@ -340,10 +374,11 @@ class Channel(PoolInstance):
         )
 
         await queue.declare(timeout=timeout)
-
         return queue
 
-    async def get_queue(self, name: str, *, ensure: bool = True) -> Queue:
+    async def get_queue(
+        self, name: str, *, ensure: bool = True
+    ) -> AbstractQueue:
         """
         With ``ensure=True``, it's a shortcut for
         ``.declare_queue(..., passive=True)``; otherwise, it returns a
@@ -365,10 +400,9 @@ class Channel(PoolInstance):
             return await self.declare_queue(name=name, passive=True)
         else:
             return self.QUEUE_CLASS(
-                connection=self,
-                channel=self.channel,
+                channel=self,
                 name=name,
-                durable=None,
+                durable=False,
                 exclusive=False,
                 auto_delete=False,
                 arguments=None,
@@ -387,12 +421,10 @@ class Channel(PoolInstance):
             warn('Use "global_" instead of "all_channels"', DeprecationWarning)
             global_ = all_channels
 
-        return await asyncio.wait_for(
-            self.channel.basic_qos(
-                prefetch_count=prefetch_count,
-                prefetch_size=prefetch_size,
-                global_=global_,
-            ),
+        return await self.channel.basic_qos(
+            prefetch_count=prefetch_count,
+            prefetch_size=prefetch_size,
+            global_=global_,
             timeout=timeout,
         )
 
@@ -404,13 +436,11 @@ class Channel(PoolInstance):
         if_empty: bool = False,
         nowait: bool = False,
     ) -> aiormq.spec.Queue.DeleteOk:
-        return await asyncio.wait_for(
-            self.channel.queue_delete(
-                queue=queue_name,
-                if_unused=if_unused,
-                if_empty=if_empty,
-                nowait=nowait,
-            ),
+        return await self.channel.queue_delete(
+            queue=queue_name,
+            if_unused=if_unused,
+            if_empty=if_empty,
+            nowait=nowait,
             timeout=timeout,
         )
 
@@ -421,10 +451,10 @@ class Channel(PoolInstance):
         if_unused: bool = False,
         nowait: bool = False,
     ) -> aiormq.spec.Exchange.DeleteOk:
-        return await asyncio.wait_for(
-            self.channel.exchange_delete(
-                exchange=exchange_name, if_unused=if_unused, nowait=nowait,
-            ),
+        return await self.channel.exchange_delete(
+            exchange=exchange_name,
+            if_unused=if_unused,
+            nowait=nowait,
             timeout=timeout,
         )
 
@@ -435,13 +465,13 @@ class Channel(PoolInstance):
                 "confirms are enabled",
             )
 
-        return Transaction(self._channel)
+        return Transaction(self)
 
     async def flow(self, active: bool = True) -> aiormq.spec.Channel.FlowOk:
         return await self.channel.flow(active=active)
 
-    def __del__(self):
+    def __del__(self) -> None:
         log.debug("%r deleted", self)
 
 
-__all__ = ("Channel", "ConfirmationTypes")
+__all__ = ("Channel",)
