@@ -1,31 +1,39 @@
+import asyncio
 from collections import defaultdict
 from logging import getLogger
-from typing import Union
+from typing import DefaultDict, Set, Type, Union
 from warnings import warn
 
 import aiormq
 
+from .abc import (
+    AbstractRobustChannel, AbstractRobustConnection, AbstractRobustExchange,
+    AbstractRobustQueue, TimeoutType,
+)
 from .channel import Channel
 from .exchange import Exchange, ExchangeType
 from .queue import Queue
 from .robust_exchange import RobustExchange
 from .robust_queue import RobustQueue
-from .tools import CallbackCollection
-from .types import TimeoutType
+from .tools import CallbackCollection, create_task
 
 
 log = getLogger(__name__)
 
 
-class RobustChannel(Channel):
+class RobustChannel(Channel, AbstractRobustChannel):
     """ Channel abstraction """
 
-    QUEUE_CLASS = RobustQueue
-    EXCHANGE_CLASS = RobustExchange
+    QUEUE_CLASS: Type[Queue] = RobustQueue
+    EXCHANGE_CLASS: Type[Exchange] = RobustExchange
+
+    _exchanges: DefaultDict[str, Set[AbstractRobustExchange]]
+    _queues: DefaultDict[str, Set[RobustQueue]]
+    default_exchange: RobustExchange
 
     def __init__(
         self,
-        connection,
+        connection: AbstractRobustConnection,
         channel_number: int = None,
         publisher_confirms: bool = True,
         on_return_raises: bool = False,
@@ -50,12 +58,13 @@ class RobustChannel(Channel):
 
         self._exchanges = defaultdict(set)
         self._queues = defaultdict(set)
-        self._prefetch_count = 0
-        self._prefetch_size = 0
-        self._global_ = False
-        self.reopen_callbacks = CallbackCollection(self)
+        self._prefetch_count: int = 0
+        self._prefetch_size: int = 0
+        self._global_qos: bool = False
+        self.reopen_callbacks: CallbackCollection = CallbackCollection(self)
 
     async def reopen(self) -> None:
+        log.debug("Reopening channel %r", self)
         await super().reopen()
         await self.restore()
         self.reopen_callbacks()
@@ -64,7 +73,7 @@ class RobustChannel(Channel):
         await self.set_qos(
             prefetch_count=self._prefetch_count,
             prefetch_size=self._prefetch_size,
-            global_=self._global_,
+            global_=self._global_qos,
         )
 
         await self.default_exchange.restore(self)
@@ -77,21 +86,19 @@ class RobustChannel(Channel):
             for queue in queues:
                 await queue.restore(self)
 
-    def _on_channel_close(self, sender, exc: BaseException):
-        if not self._is_closed_by_user:
-            self.loop.create_task(self.reopen())
+    def _on_channel_closed(self, closing: asyncio.Future) -> None:
+        super()._on_channel_closed(closing)
 
-        if exc:
-            log.exception(
-                "Robust channel %r has been closed.", sender, exc_info=exc,
-            )
-            return
+        if not self._is_closed_by_user and not self.connection.is_closed:
+            create_task(self.reopen)
+            exc = closing.exception()
+            if exc:
+                log.exception(
+                    "Robust channel %r has been closed.",
+                    self, exc_info=exc,
+                )
 
-        log.debug("Robust channel %r has been closed.", sender)
-
-    async def initialize(self, timeout: TimeoutType = None) -> None:
-        await super().initialize(timeout)
-        self.add_close_callback(self._on_channel_close)
+        log.debug("Robust channel %r has been closed.", self)
 
     async def set_qos(
         self,
@@ -100,14 +107,16 @@ class RobustChannel(Channel):
         global_: bool = False,
         timeout: TimeoutType = None,
         all_channels: bool = None,
-    ):
+    ) -> aiormq.spec.Basic.QosOk:
         if all_channels is not None:
             warn('Use "global_" instead of "all_channels"', DeprecationWarning)
             global_ = all_channels
 
+        await self.connection.connected.wait()
+
         self._prefetch_count = prefetch_count
         self._prefetch_size = prefetch_size
-        self._global_ = global_
+        self._global_qos = global_
 
         return await super().set_qos(
             prefetch_count=prefetch_count,
@@ -120,38 +129,42 @@ class RobustChannel(Channel):
         self,
         name: str,
         type: Union[ExchangeType, str] = ExchangeType.DIRECT,
-        durable: bool = None,
+        durable: bool = False,
         auto_delete: bool = False,
         internal: bool = False,
         passive: bool = False,
         arguments: dict = None,
         timeout: TimeoutType = None,
         robust: bool = True,
-    ) -> Exchange:
-        exchange = await super().declare_exchange(
-            name=name,
-            type=type,
-            durable=durable,
-            auto_delete=auto_delete,
-            internal=internal,
-            passive=passive,
-            arguments=arguments,
-            timeout=timeout,
+    ) -> AbstractRobustExchange:
+        await self.connection.connected.wait()
+        exchange = (
+            await super().declare_exchange(
+                name=name,
+                type=type,
+                durable=durable,
+                auto_delete=auto_delete,
+                internal=internal,
+                passive=passive,
+                arguments=arguments,
+                timeout=timeout,
+            )
         )
 
         if not internal and robust:
-            self._exchanges[name].add(exchange)
+            # noinspection PyTypeChecker
+            self._exchanges[name].add(exchange)     # type: ignore
 
-        return exchange
+        return exchange     # type: ignore
 
     async def exchange_delete(
         self,
         exchange_name: str,
         timeout: TimeoutType = None,
-        if_unused=False,
-        nowait=False,
+        if_unused: bool = False,
+        nowait: bool = False,
     ) -> aiormq.spec.Exchange.DeleteOk:
-
+        await self.connection.connected.wait()
         result = await super().exchange_delete(
             exchange_name=exchange_name,
             timeout=timeout,
@@ -167,16 +180,16 @@ class RobustChannel(Channel):
         self,
         name: str = None,
         *,
-        durable: bool = None,
+        durable: bool = False,
         exclusive: bool = False,
         passive: bool = False,
         auto_delete: bool = False,
         arguments: dict = None,
         timeout: TimeoutType = None,
         robust: bool = True
-    ) -> Queue:
-
-        queue = await super().declare_queue(
+    ) -> AbstractRobustQueue:
+        await self.connection.connected.wait()
+        queue: RobustQueue = await super().declare_queue(   # type: ignore
             name=name,
             durable=durable,
             exclusive=exclusive,
@@ -187,7 +200,7 @@ class RobustChannel(Channel):
         )
 
         if robust:
-            self._queues[name].add(queue)
+            self._queues[queue.name].add(queue)
 
         return queue
 
@@ -198,7 +211,8 @@ class RobustChannel(Channel):
         if_unused: bool = False,
         if_empty: bool = False,
         nowait: bool = False,
-    ):
+    ) -> aiormq.spec.Queue.DeleteOk:
+        await self.connection.connected.wait()
         result = await super().queue_delete(
             queue_name=queue_name,
             timeout=timeout,
