@@ -1,112 +1,111 @@
 import asyncio
-from functools import wraps
 from logging import getLogger
-from typing import Callable, Type
+from typing import Any, Callable, Dict, Optional, Type, Union
+from warnings import warn
 from weakref import WeakSet
 
+from aiormq.abc import ExceptionType
 from aiormq.connection import parse_bool, parse_int
+from pamqp.common import FieldTable
+from yarl import URL
 
-from .connection import Connection, ConnectionType, connect
+from .abc import (
+    AbstractChannel, AbstractConnection, AbstractRobustChannel,
+    AbstractRobustConnection, TimeoutType,
+)
+from .connection import Connection, make_url
 from .exceptions import CONNECTION_EXCEPTIONS
 from .robust_channel import RobustChannel
-from .tools import CallbackCollection
-from .types import TimeoutType
+from .tools import CallbackCollection, task
 
 
 log = getLogger(__name__)
 
 
-def _ensure_connection(func):
-    @wraps(func)
-    def wrap(self, *args, **kwargs):
-        if self.is_closed:
-            raise RuntimeError("Connection closed")
-
-        return func(self, *args, **kwargs)
-
-    return wrap
-
-
-class RobustConnection(Connection):
+class RobustConnection(Connection, AbstractRobustConnection):
     """ Robust connection """
 
-    CHANNEL_CLASS = RobustChannel
+    CHANNEL_REOPEN_PAUSE = 1
+    CHANNEL_CLASS: Type[RobustChannel] = RobustChannel
     KWARGS_TYPES = (
         ("reconnect_interval", parse_int, "5"),
         ("fail_fast", parse_bool, "1"),
     )
 
-    def __init__(self, url, loop=None, **kwargs):
+    def __init__(
+        self, url: URL, loop: asyncio.AbstractEventLoop = None, **kwargs: Any
+    ):
         super().__init__(url=url, loop=loop, **kwargs)
 
-        self.connect_kwargs = {}
+        self.connect_kwargs: Dict[str, Any] = {}
         self.reconnect_interval = self.kwargs["reconnect_interval"]
         self.fail_fast = self.kwargs["fail_fast"]
 
-        self.__channels = WeakSet()
-        self._reconnect_callbacks = CallbackCollection(self)
+        self.__channels: WeakSet[AbstractChannel] = WeakSet()
         self._connect_lock = asyncio.Lock()
-        self._closed = False
-        self.connected = asyncio.Event()
+        self._is_closed_by_user = False
+        self.reconnect_callbacks: CallbackCollection = CallbackCollection(self)
 
     @property
     def reconnecting(self) -> bool:
         return self._connect_lock.locked()
 
-    @property
-    def reconnect_callbacks(self) -> CallbackCollection:
-        return self._reconnect_callbacks
-
-    @property
-    def _channels(self) -> dict:
-        return {ch.number: ch for ch in self.__channels}
-
-    def __repr__(self):
+    def __repr__(self) -> str:
         return '<{0}: "{1}" {2} channels>'.format(
             self.__class__.__name__, str(self), len(self.__channels),
         )
 
-    def _on_connection_close(self, connection, closing, *args, **kwargs):
+    def _on_connection_close(
+        self, connection: AbstractConnection, closing: asyncio.Future,
+    ) -> None:
         if self.reconnecting:
             return
 
         self.connected.clear()
-        self.connection = None
+        del self.connection
 
-        super()._on_connection_close(connection, closing)
+        log.debug("Closing AMQP connection %r", connection)
 
-        if self._closed:
+        if self.closing.done():
             return
 
         log.info(
             "Connection to %s closed. Reconnecting after %r seconds.",
-            self,
-            self.reconnect_interval,
+            self, self.reconnect_interval,
         )
         self.loop.call_later(
             self.reconnect_interval,
-            lambda: self.loop.create_task(self.reconnect()),
+            self.reconnect,
         )
 
     def add_reconnect_callback(
-        self, callback: Callable[[], None], weak: bool = False
-    ):
+        self, callback: Callable[["RobustConnection"], None],
+        weak: bool = False,
+    ) -> None:
         """ Add callback which will be called after reconnect.
 
         :return: None
         """
+        warn(
+            "This method will be removed from future release. "
+            f"Use {self.__class__.__name__}.reconnect_callbacks.add instead",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        self.reconnect_callbacks.add(callback, weak=weak)
 
-        self._reconnect_callbacks.add(callback, weak=weak)
-
-    async def __cleanup_connection(self, exc):
-        if self.connection is None:
+    async def __cleanup_connection(self, exc: Optional[BaseException]) -> None:
+        if not hasattr(self, "connection"):
             return
+
         await asyncio.gather(
             self.connection.close(exc), return_exceptions=True,
         )
-        self.connection = None
+        del self.connection
 
-    async def connect(self, timeout: TimeoutType = None, **kwargs):
+    async def connect(
+        self, timeout: TimeoutType = None, **kwargs: Any
+    ) -> None:
         if self.is_closed:
             raise RuntimeError("{!r} connection closed".format(self))
 
@@ -115,25 +114,24 @@ class RobustConnection(Connection):
             self.connect_kwargs = kwargs
 
         if self.reconnecting:
-            log.warning(
-                "Connect method called but connection %r is "
-                "reconnecting right now.",
-                self,
+            raise RuntimeError(
+                (
+                    f"Connect method called but connection "
+                    f"{self!r} is reconnecting right now."
+                ), self,
             )
 
         async with self._connect_lock:
-            while True:
+            while not self.closing.done():
                 try:
-                    result = await super().connect(
+                    await super().connect(
                         timeout=timeout, **self.connect_kwargs
                     )
 
-                    for channel in self._channels.values():
+                    for channel in self.__channels:
                         await channel.reopen()
 
                     self.fail_fast = False
-                    self.connected.set()
-                    return result
                 except CONNECTION_EXCEPTIONS as e:
                     if self.fail_fast:
                         raise
@@ -150,48 +148,64 @@ class RobustConnection(Connection):
                 except asyncio.CancelledError as e:
                     await self.__cleanup_connection(e)
                     raise
+                else:
+                    self.connected.set()
+                    return
 
+                log.info(
+                    "Reconnect attempt failed %s. Retrying after %r seconds.",
+                    self, self.reconnect_interval,
+                )
                 await asyncio.sleep(self.reconnect_interval)
 
-    async def reconnect(self):
+    @task
+    async def reconnect(self) -> None:
         await self.connect()
-        self._reconnect_callbacks(self)
+        self.reconnect_callbacks(self)
 
     def channel(
         self,
         channel_number: int = None,
         publisher_confirms: bool = True,
-        on_return_raises=False,
-    ):
+        on_return_raises: bool = False,
+    ) -> AbstractRobustChannel:
 
-        channel = super().channel(
+        channel: AbstractRobustChannel = super().channel(
             channel_number=channel_number,
             publisher_confirms=publisher_confirms,
             on_return_raises=on_return_raises,
-        )
+        )   # type: ignore
 
         self.__channels.add(channel)
+        self.close_callbacks.add(
+            lambda c, e: channel.close_callbacks(e),
+        )
 
         return channel
 
     @property
-    def is_closed(self):
+    def is_closed(self) -> bool:
         """ Is this connection is closed """
-        return self._closed or super().is_closed
+        return self._is_closed_by_user or super().is_closed
 
-    async def close(self, exc=asyncio.CancelledError):
+    async def close(
+        self, exc: Optional[ExceptionType] = asyncio.CancelledError,
+    ) -> None:
         if self.is_closed:
             return
 
-        self._closed = True
+        self._is_closed_by_user = True
 
-        if self.connection is None:
+        if not hasattr(self, "connection"):
             return
-        return await super().close(exc)
+
+        result = await super().close(exc)
+        self.close_callbacks(exc)
+        return result
 
 
 async def connect_robust(
-    url: str = None,
+    url: Union[str, URL] = None,
     *,
     host: str = "localhost",
     port: int = 5672,
@@ -202,16 +216,12 @@ async def connect_robust(
     loop: asyncio.AbstractEventLoop = None,
     ssl_options: dict = None,
     timeout: TimeoutType = None,
-    connection_class: Type[ConnectionType] = RobustConnection,
-    client_properties: dict = None,
-    **kwargs
-) -> ConnectionType:
+    client_properties: FieldTable = None,
+    connection_class: Type[AbstractRobustConnection] = RobustConnection,
+    **kwargs: Any
+) -> AbstractRobustConnection:
 
-    """ Make robust connection to the broker.
-
-    That means that connection state will be restored after reconnect.
-    After connection has been established the channels, the queues and the
-    exchanges with their bindings will be restored.
+    """ Make connection to the broker.
 
     Example:
 
@@ -220,7 +230,7 @@ async def connect_robust(
         import aio_pika
 
         async def main():
-            connection = await aio_pika.connect_robust(
+            connection = await aio_pika.connect(
                 "amqp://guest:guest@127.0.0.1/"
             )
 
@@ -231,7 +241,7 @@ async def connect_robust(
         import aio_pika
 
         async def main():
-            connection = await aio_pika.connect_robust()
+            connection = await aio_pika.connect()
 
     .. note::
 
@@ -244,9 +254,30 @@ async def connect_robust(
         For an information on what the ssl_options can be set to reference the
         `official Python documentation`_ .
 
+    Set connection name for RabbitMQ admin panel:
+
+    .. code-block:: python
+
+        read_connection = await connect(
+            client_properties={
+                'connection_name': 'Read connection'
+            }
+        )
+
+        write_connection = await connect(
+            client_properties={
+                'connection_name': 'Write connection'
+            }
+        )
+
+    .. note:
+
+        ``client_properties`` argument requires ``aiormq>=2.9``
+
     URL string might be contain ssl parameters e.g.
     `amqps://user:pass@host//?ca_certs=ca.pem&certfile=crt.pem&keyfile=key.pem`
 
+    :param client_properties: add custom client capability.
     :param url:
         RFC3986_ formatted broker address. When :class:`None`
         will be used keyword arguments.
@@ -267,22 +298,31 @@ async def connect_robust(
     .. _RFC3986: https://goo.gl/MzgYAs
     .. _official Python documentation: https://goo.gl/pty9xA
 
+
     """
-    return await connect(
-        url=url,
-        host=host,
-        port=port,
-        login=login,
-        password=password,
-        virtualhost=virtualhost,
-        ssl=ssl,
+
+    connection: AbstractRobustConnection = connection_class(
+        make_url(
+            url,
+            host=host,
+            port=port,
+            login=login,
+            password=password,
+            virtualhost=virtualhost,
+            ssl=ssl,
+            ssl_options=ssl_options,
+            **kwargs
+        ),
         loop=loop,
-        connection_class=connection_class,
-        ssl_options=ssl_options,
+    )
+
+    await connection.connect(
         timeout=timeout,
         client_properties=client_properties,
-        **kwargs
+        loop=loop,
     )
+
+    return connection
 
 
 __all__ = (
