@@ -1,46 +1,47 @@
-import os
-from base64 import b32encode
-from collections import namedtuple
-from logging import getLogger
-from types import FunctionType
+import logging
+from random import getrandbits
+from typing import Any, Callable, Dict, Optional, Tuple, Union
 
 import aiormq
+from pamqp.common import Arguments
 
-from .channel import Channel
-from .exchange import ExchangeParamType
-from .queue import ConsumerTag, Queue
-
-
-log = getLogger(__name__)
-
-
-DeclarationResult = namedtuple(
-    "DeclarationResult", ("message_count", "consumer_count"),
+# This needed only for migration from 6.x to 7.x
+# TODO: Remove this in 8.x release
+from .abc import DeclarationResult  # noqa
+from .abc import (
+    AbstractChannel, AbstractExchange, AbstractIncomingMessage,
+    AbstractRobustChannel, AbstractRobustQueue, ConsumerTag, TimeoutType,
 )
+from .exchange import ExchangeParamType
+from .queue import Queue
 
 
-class RobustQueue(Queue):
+log = logging.getLogger(__name__)
+
+
+class RobustQueue(Queue, AbstractRobustQueue):
     __slots__ = ("_consumers", "_bindings")
 
+    _consumers: Dict[ConsumerTag, Dict[str, Any]]
+    _bindings: Dict[Tuple[Union[AbstractExchange, str], str], Dict[str, Any]]
+
     @staticmethod
-    def _get_random_queue_name():
-        rb = os.urandom(16)
-        return "amq_%s" % b32encode(rb).decode().replace("=", "").lower()
+    def _get_random_queue_name() -> str:
+        rnd = getrandbits(128)
+        return "amq_%s" % hex(rnd).lower()
 
     def __init__(
         self,
-        connection,
-        channel: aiormq.Channel,
-        name,
-        durable,
-        exclusive,
-        auto_delete,
-        arguments,
+        channel: AbstractChannel,
+        name: Optional[str],
+        durable: bool = False,
+        exclusive: bool = False,
+        auto_delete: bool = False,
+        arguments: Arguments = None,
         passive: bool = False,
     ):
 
         super().__init__(
-            connection=connection,
             channel=channel,
             name=name or self._get_random_queue_name(),
             durable=durable,
@@ -53,8 +54,8 @@ class RobustQueue(Queue):
         self._consumers = {}
         self._bindings = {}
 
-    async def restore(self, channel: Channel):
-        self._channel = channel._channel
+    async def restore(self, channel: AbstractRobustChannel) -> None:
+        self.channel = channel
 
         await self.declare()
 
@@ -69,22 +70,23 @@ class RobustQueue(Queue):
         exchange: ExchangeParamType,
         routing_key: str = None,
         *,
-        arguments=None,
-        timeout: int = None,
+        arguments: Arguments = None,
+        timeout: TimeoutType = None,
         robust: bool = True
-    ):
-
+    ) -> aiormq.spec.Queue.BindOk:
+        await self.connection.connected.wait()
         if routing_key is None:
             routing_key = self.name
 
-        kwargs = dict(arguments=arguments, timeout=timeout)
-
         result = await super().bind(
-            exchange=exchange, routing_key=routing_key, **kwargs
+            exchange=exchange, routing_key=routing_key,
+            arguments=arguments, timeout=timeout,
         )
 
         if robust:
-            self._bindings[(exchange, routing_key)] = kwargs
+            self._bindings[(exchange, routing_key)] = dict(
+                arguments=arguments,
+            )
 
         return result
 
@@ -92,10 +94,10 @@ class RobustQueue(Queue):
         self,
         exchange: ExchangeParamType,
         routing_key: str = None,
-        arguments: dict = None,
-        timeout: int = None,
-    ):
-
+        arguments: Arguments = None,
+        timeout: TimeoutType = None,
+    ) -> aiormq.spec.Queue.UnbindOk:
+        await self.connection.connected.wait()
         if routing_key is None:
             routing_key = self.name
 
@@ -108,35 +110,41 @@ class RobustQueue(Queue):
 
     async def consume(
         self,
-        callback: FunctionType,
+        callback: Callable[[AbstractIncomingMessage], Any],
         no_ack: bool = False,
         exclusive: bool = False,
-        arguments: dict = None,
-        consumer_tag=None,
-        timeout=None,
+        arguments: Arguments = None,
+        consumer_tag: ConsumerTag = None,
+        timeout: TimeoutType = None,
         robust: bool = True,
     ) -> ConsumerTag:
-
-        kwargs = dict(
+        await self.connection.connected.wait()
+        consumer_tag = await super().consume(
+            consumer_tag=consumer_tag,
+            timeout=timeout,
             callback=callback,
             no_ack=no_ack,
             exclusive=exclusive,
             arguments=arguments,
         )
 
-        consumer_tag = await super().consume(
-            consumer_tag=consumer_tag, **kwargs
-        )
-
         if robust:
-            self._consumers[consumer_tag] = kwargs
+            self._consumers[consumer_tag] = dict(
+                callback=callback,
+                no_ack=no_ack,
+                exclusive=exclusive,
+                arguments=arguments,
+            )
 
         return consumer_tag
 
     async def cancel(
-        self, consumer_tag: ConsumerTag, timeout=None, nowait: bool = False
-    ):
-
+        self,
+        consumer_tag: ConsumerTag,
+        timeout: TimeoutType = None,
+        nowait: bool = False,
+    ) -> aiormq.spec.Basic.CancelOk:
+        await self.connection.connected.wait()
         result = await super().cancel(consumer_tag, timeout, nowait)
         self._consumers.pop(consumer_tag, None)
         return result
