@@ -9,15 +9,13 @@ import aiormq
 import pamqp
 import pytest
 from aiomisc import awaitable
-from async_generator import async_generator, yield_
+from aiormq.connection import DEFAULT_PORTS
 from yarl import URL
 
 import aio_pika
-from aiormq.connection import DEFAULT_PORTS
 
 
 @pytest.fixture
-@async_generator
 async def add_cleanup(loop):
     entities = []
 
@@ -27,7 +25,7 @@ async def add_cleanup(loop):
         entities.append(func)
 
     try:
-        await yield_(payload)
+        yield payload
     finally:
         for func in entities[::-1]:
             await func()
@@ -36,7 +34,6 @@ async def add_cleanup(loop):
 
 
 @pytest.fixture
-@async_generator
 async def create_task(loop):
     tasks = []
 
@@ -47,7 +44,7 @@ async def create_task(loop):
         return task
 
     try:
-        await yield_(payload)
+        yield payload
     finally:
         cancelled = []
         for task in tasks:
@@ -78,8 +75,10 @@ def amqp_direct_url(request) -> URL:
 
 
 @pytest.fixture
-def amqp_url(amqp_direct_url) -> URL:
-    return amqp_direct_url
+def amqp_url(request, amqp_direct_url) -> URL:
+    query = dict(amqp_direct_url.query)
+    query["name"] = request.node.nodeid
+    return amqp_direct_url.with_query(**query)
 
 
 @pytest.fixture(
@@ -117,18 +116,16 @@ def create_channel(connection: aio_pika.Connection, add_cleanup):
 
 # noinspection PyTypeChecker
 @pytest.fixture
-@async_generator
 async def connection(create_connection) -> aio_pika.Connection:
     async with await create_connection() as conn:
-        await yield_(conn)
+        yield conn
 
 
 # noinspection PyTypeChecker
 @pytest.fixture
-@async_generator
 async def channel(connection: aio_pika.Connection) -> aio_pika.Channel:
     async with connection.channel() as ch:
-        await yield_(ch)
+        yield ch
 
 
 @pytest.fixture
