@@ -5,6 +5,7 @@ from functools import partial
 import pytest
 
 import aio_pika
+from aio_pika import RobustChannel
 from tests.test_amqp import (
     TestCaseAmqp, TestCaseAmqpNoConfirms, TestCaseAmqpWithConfirms,
 )
@@ -48,24 +49,24 @@ class TestCaseNoRobust(TestCaseAmqp):
         def cb(*a, **kw):
             pass
 
-        connection.add_reconnect_callback(cb)
+        connection.reconnect_callbacks.add(cb)
 
         del cb
         assert len(connection.reconnect_callbacks) == 1
 
     async def test_channel_blocking_timeout_reopen(self, connection):
-        channel = await connection.channel()
+        channel: RobustChannel = await connection.channel()     # type: ignore
         close_reasons = []
         close_event = asyncio.Event()
         reopen_event = asyncio.Event()
-        channel.reopen_callbacks.add(lambda _: reopen_event.set(), weak=False)
+        channel.reopen_callbacks.add(lambda _: reopen_event.set())
 
         def on_done(*args):
             close_reasons.append(args)
             close_event.set()
             return
 
-        channel.add_close_callback(on_done)
+        channel.close_callbacks.add(on_done)
 
         async def run(sleep_time=1):
             await channel.set_qos(1)
